Cleanup: avoid warnings for mc-library on x86_64

+++ lib/tty/color-internal.h
@@ -24,7 +24,7 @@ typedef struct mc_color_pair_struct {
     const char *cbg;
     int ifg;
     int ibg;
-    int pair_index;
+    size_t pair_index;
     gboolean is_temp;
 } tty_color_pair_t;
